+++ package.json
@@ -55,6 +55,7 @@
     "eslint-config-next": "^13.4.19",
     "eslint-plugin-import": "^2.29.1",
     "eslint-plugin-jest-formatting": "^3.1.0",
+    "eslint-plugin-unused-imports": "^3.0.0",
     "jsdom": "^22.1.0",
     "postcss": "^8.4.27",
     "prettier": "^3.0.0",
