+++ src/features/auth/AuthShowcase.tsx
@@ -1,0 +1,19 @@
+import { signIn, signOut, useSession } from "next-auth/react";
+
+export function AuthShowcase() {
+  const { data: sessionData } = useSession();
+
+  return (
+    <div className="flex flex-col items-center justify-center gap-4 ">
+      <p className="text-end text-xl">
+        {sessionData && <span>Logged in as {sessionData.user?.name}</span>}
+      </p>
+      <button
+        className="rounded-full bg-white/10 px-10 py-3 font-semibold  no-underline transition hover:bg-white/20"
+        onClick={sessionData ? () => void signOut() : () => void signIn()}
+      >
+        {sessionData ? "Sign out" : "Sign in"}
+      </button>
+    </div>
+  );
+}
