+++ src/pages/space/[id]/join/index.tsx
@@ -1,7 +1,6 @@
 import { useSession } from "next-auth/react";
 import { useRouter } from "next/router";
 import { signIn } from "next-auth/react";
-import { toast } from "sonner";
 
 import { Loader } from "~/features/common/components";
 import { Container } from "~/features/common/components/layout";
